apps/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg[binary]>=3.2
python-dotenv>=1.0

# 선택
# alembic>=1.13